# docs/swagger.yaml
basePath: /
definitions:
...
  internal_organization.AddMemberRequest:
    properties:
      role:
        enum:
        - administrator
        - participant
        - user
        example: participant
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - user_id
    type: object
  internal_organization.CreateRequest:
    properties:
      name:
        example: My Company
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  internal_organization.CreateResponse:
    properties:
      data:
        properties:
          avatar_url:
            type: string
          created_at:
            type: string
          id:
            type: string
          name:
            type: string
          owner_id:
            type: string
          slug:
            type: string
          status:
            type: string
          updated_at:
            type: string
        type: object
    type: object
  internal_organization.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_organization.ListResponse:
    properties:
      data:
        properties:
          organizations:
            items:
              properties:
                avatar_url:
                  type: string
                created_at:
                  type: string
                id:
                  type: string
                name:
                  type: string
                owner_id:
                  type: string
                slug:
                  type: string
                status:
                  type: string
                updated_at:
                  type: string
              type: object
            type: array
        type: object
    type: object
  internal_organization.UpdateRequest:
    properties:
      name:
        example: My Company Updated
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API для работы с командой над проектом.
  title: timeManager API
  version: "1.0"
paths:
...
  /api/v1/organizations:
    get:
      description: Возвращает организации, в которых текущий пользователь является
        участником.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_organization.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список моих организаций
      tags:
      - organizations
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Создаёт организацию от имени текущего пользователя. Slug генерируется
        автоматически из name. Поддерживает JSON или multipart/form-data (name + опционально
        avatar).
      parameters:
      - description: 'Название организации (JSON) или multipart: name, avatar'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_organization.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_organization.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать организацию
      tags:
      - organizations
  /api/v1/organizations/{id}:
    get:
      description: Возвращает организацию. Доступ только для членов организации.
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить организацию по ID
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Обновляет название и опционально аватарку. Slug остаётся неизменным.
        Поддерживает JSON или multipart (name + опционально avatar).
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      - description: 'Новое название (JSON) или multipart: name, avatar'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_organization.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_organization.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить организацию
      tags:
      - organizations
  /api/v1/organizations/{id}/archive:
    post:
      consumes:
      - application/json
      description: Переводит организацию в статус archived.
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_organization.CreateResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Архивировать организацию
      tags:
      - organizations
  /api/v1/organizations/{id}/avatar:
    put:
      consumes:
      - multipart/form-data
      description: Загружает аватарку в MinIO. multipart/form-data с полем avatar
        (jpeg/png/webp/gif, max 5MB).
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      - description: Файл изображения
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_organization.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Установить аватарку организации
      tags:
      - organizations
  /api/v1/organizations/{id}/members:
    get:
      description: Возвращает всех пользователей, входящих в организацию. Доступ только
        для членов организации.
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список участников организации
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Добавляет пользователя в организацию. Пользователь может быть только
        в одной организации.
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_organization.AddMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить участника в организацию
      tags:
      - organizations
  /api/v1/organizations/{id}/members/{user_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить участника из организации
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"